move files to own folders
