+++ src/components/Button/HeartButton/HeartButton.jsx
@@ -1,5 +1,5 @@
 import { HeartBtn } from "./HeartButton.styled";
-import Icon from "./icons/index";
+import Icon from "../icons/index";
 
 
 export const HeartButton = () => {
