+++ src/components/Button/TrashButton/TrashButton.styled.js
