+++ src/components/Button/icons/index.js
@@ -1,7 +1,9 @@
 import { ReactComponent as Heart } from "./heart.svg";
 import { ReactComponent as Trash } from "./trash.svg";
 
-export default {
+const Icon = {
   Heart,
   Trash,
 };
+
+export default Icon;
