+++ src/components/Button/index.js
@@ -1,5 +1,5 @@
-export { LearnMoreButton } from "./LearnMoreButton";
-export { HeartButton } from "./HeartButton";
-export { TrashButton } from "./TrashButton";
-export { CancelButton } from "./CancelButton";
-export { NextButton } from "./NextButton";
+export { LearnMoreButton } from "./LearnMoreButton/LearnMoreButton";
+export { HeartButton } from "./HeartButton/HeartButton";
+export { TrashButton } from "./TrashButton/TrashButton";
+export { CancelButton } from "./CancelButton/CancelButton";
+export { NextButton } from "./NextButton/NextButton";
